refactor(frontend): consolidate request context into apiFetch module
Merge requestContext functionality into apiFetch.server.ts:
- Move RequestContext interface and all related functions into apiFetch.server.ts
- Move AsyncLocalStorage and helper functions (generateRandomHex, initializeRequestContext, etc.)
- Delete app/utils/requestContext.server.ts
- Reduce number of files by consolidating related functionality

This simplifies the codebase by keeping request context logic together with its primary consumer.
